feat: 파일 구조 변경 및 프리셋 추가
