+++ example/lib/main.dart
@@ -59,96 +59,6 @@ class _HomeTabState extends State<HomeTab> with SingleTickerProviderStateMixin {
     super.dispose();
   }
 
-  Widget customBuilder(
-    BuildContext context,
-    PageController pageController,
-    List<Widget> childrenWithKey,
-    int index,
-  ) {
-    return AnimatedBuilder(
-      animation: pageController,
-      builder: (context, child) {
-        double page = 0.0;
-        try {
-          page = pageController.page ?? pageController.initialPage.toDouble();
-        } catch (_) {
-          page = pageController.initialPage.toDouble();
-        }
-        double offset = page - index;
-        // Custom animation: outgoing page moves 0~20% and fades out,
-        // incoming page moves 80~100% and fades in.
-        double dx = 0.0;
-        double opacity = 1.0;
-        final width = MediaQuery.of(context).size.width;
-
-        // Outgoing page (current)
-        if (offset > 0 && offset <= 1) {
-          // Swiping right, this page is outgoing
-          if (offset > 0.5) {
-            // 0.5~1.0: move 80%~100%, opacity 0~1
-            double t = (offset - 0.5) / 0.5;
-            dx = (0.6 + t * 0.4) * width;
-            opacity = t;
-          } else {
-            // 0~0.5: move stays at 80%, opacity 0
-            dx = 0.6 * width;
-            opacity = 0.0;
-          }
-        } else if (offset < 0 && offset >= -1) {
-          // Swiping left, this page is outgoing
-          if (offset < -0.5) {
-            // -0.5~-1.0: move -80%~-100%, opacity 0~1
-            double t = (-offset - 0.5) / 0.5;
-            dx = -(0.6 + t * 0.4) * width;
-            opacity = t;
-          } else {
-            // 0~ -0.5: move stays at -80%, opacity 0
-            dx = -0.6 * width;
-            opacity = 0.0;
-          }
-        } else if (offset == 0) {
-          dx = 0;
-          opacity = 1.0;
-        }
-
-        // Incoming page
-        if (offset > 0 && offset <= 1) {
-          // This page is incoming from left
-          if (offset <= 0.5) {
-            // 0~0.5: move 0~20%, fade 1~0
-            double t = offset / 0.5;
-            dx = t * 0.4 * width;
-            opacity = 1.0 - t;
-          } else {
-            // 0.5~1.0: move stays at 20%, opacity 0
-            dx = 0.4 * width;
-            opacity = 0.0;
-          }
-        } else if (offset < 0 && offset >= -1) {
-          // This page is incoming from right
-          if (offset >= -0.5) {
-            // 0~ -0.5: move 0~ -20%, fade 1~0
-            double t = -offset / 0.5;
-            dx = -t * 0.4 * width;
-            opacity = 1.0 - t;
-          } else {
-            // -0.5~ -1.0: move stays at -20%, opacity 0
-            dx = -0.4 * width;
-            opacity = 0.0;
-          }
-        }
-        return Opacity(
-          opacity: opacity.clamp(0.0, 1.0),
-          child: Transform.scale(
-            scale: offset.abs() < 1 ? 1 - offset.abs() * 0.1 : 0.8,
-            child: Transform.translate(offset: Offset(dx, 0), child: child),
-          ),
-        );
-      },
-      child: childrenWithKey[index],
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -156,13 +66,12 @@ class _HomeTabState extends State<HomeTab> with SingleTickerProviderStateMixin {
         title: const Text('CustomTabBarView Demo'),
       ),
       backgroundColor: Colors.white,
-      body: CustomTabBarView.builder(
+      body: CustomTabBarView.stack(
         controller: _tabController,
         physics: const PageScrollPhysics(),
         dragStartBehavior: DragStartBehavior.down,
         // children: _children,
         tabs: _children,
-        builder: customBuilder,
       ),
       bottomNavigationBar: SafeArea(
         child: Transform.flip(
