+++ lib/custom_tabbarview.dart
@@ -1,3 +1,4 @@
 library custom_tabbarview;
 
 export 'src/custom_tabbarview.dart';
+export 'src/custom_tabbarview_types.dart';
