+++ lib/src/custom_tabbarview_types.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+typedef CustomTabBarViewBuilder = Widget Function(
+  BuildContext context,
+  PageController pageController,
+  List<Widget> childrenWithKey,
+  int index,
+);
